general refactor with ruff
