+++ packages/models-library/src/models_library/api_schemas_directorv2/dynamic_services_service.py
@@ -4,10 +4,11 @@ from typing import Any, ClassVar
 
 from pydantic import BaseModel, Field
 
+from ..basic_regex import VERSION_RE
 from ..basic_types import PortInt
 from ..projects import ProjectID
 from ..projects_nodes_io import NodeID
-from ..services import VERSION_RE, DynamicServiceKey
+from ..services import DynamicServiceKey
 from ..services_enums import ServiceBootType, ServiceState
 from ..users import UserID
